Partial refactor to simplify (hopefully) the preheat/roast models.
Ideally, the timer service doesn't have to be so aware of the roast
state and preheat vs roast timers etc., and ideally all durations in the
roast logs can be relative to a universal start time (most sensibly? the
very start of the roast, whether that was a preheat or not).

Introduce a RoastTimeline model that can serve as a sort of view model
for most of the app. Many services are sort of repeating work as they
comb through the roast logs (raw or from RoastLogProvider) to find
phases etc.

The RoastTimeline is not plugged in universally yet and it has its own
strangeness TBD. Should PhaseLog even exist? It is not typesafe in that
it allows for nonsense like multiple START times, or an END time without
a START.

The relationship between RoastState and RoastTimeline is also somewhat
bidirectional. Consider preheat; should preheat time be a Duration set
to zero that allows the state to be inferred as preheating; or should
the state be a separate member so that the timeline can be set to
preheating without having to add such a useless Duration to it?

How exactly should RoastTimeline and TimerService interact, especially
in terms of the differences between preheat timer and roast timer and
the checktemp intervals that come? How to best minimize the fragility
that comes with having two different concepts of the current roast time?

Finally, this has a bug in CheckTempWidget, and, the roast logs widgets
display of preheat timing isn't super useful as it shows the start time
in negatives relative to when the preheat ended making it hard to tell
how long the preheat stage actually was etc.

Didn't expect preheat to be so complicated! (It's probably not I just
mucked up the order of refactoring to integrate it more gracefully).