Fix bug, preheat not showing 0:00 in TimeWidget.
Subtle issue. There are no listeners to preheatSecondsProvider in the
preheat screen with the start button. When you hit start, the roast
state is updated but the widget tree is not rebuilt until after the
timer is started and 0:00 is added to the stream. After this the
onPressed handler completes and the widget tree rebuilds, and the new UI
watches preheatSecondsProvider. However, since the seconds stream was a
broadcast() stream controller, it means the 0:00 event got dropped.

This was not happening for roast time, because roastSecondsProvider was
already being watched, and so it was constructed and the 0:00 event for
the roast was created while the stream was being observed by riverpod.

I don't think there's a reason to use broadcast here; removing that
means the event sticks around and is consumed when the new UI loads.