New feature: edit previously recorded temperatures.
Note, doesn't allow for editing phase logs, or control logs, or for
editing timestamps, or for deleting logs. Some or all of these may be
important?

Editing/deleting phase logs would be particularly thorny (eg, deleting
dryEnd after logging first crack would put the app in a bad state).

Additionally, its not clear that the ability to edit more than just the
most recent temp log is sufficiently discoverable. Hopefully this need
is rare, and the discoverability will be good enough. But without users,
it is certainly hard to say for sure!