Change default preheat temp to 150, reachable in recommended time.
While the internet recommends preheating to 200, Behmor does not
recommend preheating more than 1:45. That time allows us to reach about
155.

We show a warning at 1:45, and therefore it's to some extent a poor trap
to set users up for a target preheat temp that triggers a warning. While
many users will likely choose to exceed this guideline, let's not
recommend a temp that requires them to do so.