Refine smoke suppressor "on" time, add an alert for when it turns off.
A reasonable profile will involve selecting the highest power setting
before the smoke suppressor turns on, to prevent the roast temps from
dropping. How long the smoke suppressor stays on is then unclear, and
if you maintain a high temp (say, 320 degrees), when it turns off you
might quickly overheat. This is especially risky because of lag time
when you reduce power to the heating element.

Send the second alert only if we're _reasonably_ within striking
distance of max temps, and if we're on the highest power level.