Add new RoastManagerService to manage roast lifetime.
This functions more like a BLoC than a riverpod state object. The goal
is that this makes it easier to efficiently cache and recompute
thermodynamics calculations, so that we can add them to live roasting
without bogging things down and skipping frames.

This also nicely moves a lot of roast logic out of views which is good.

The only pain points of this design are: having `.valueOrNull`,
`.requireValue` everywhere (because these streams are an abstraction,
and not truly asynchronous, so we can expect values)...and, the fact
that Riverpod doesn't initialize them synchronously, so we have to
preload...and, the fact that it doesn't necessarily integrate perfectly
at the boundary with riverpod.