Fix SafeArea usage (blocked by Scaffolds with bottomApBar).
Previous SafeArea's weren't all working because they rely on
MediaQuery.of(context).padding, which is reset in Scaffold.body if a
bottomAppBar is specified.

The BottomAppBar that was being used for preheat page, had no actual
contents. But because it wasn't null, it still meant the body had no
padding to use in SafeArea.

The same area occurs with the `toast` from TimerScaffold (the roast tips
widget on the inner of the two scaffolds, used to properly position the
floating action buttons). Therefore, just wrap the inner scaffold in a
SafeArea.

Still requires the bottomAppBar to take care of the actual padding,
which was handled in the previous commit by the PhaseControlsWidget.

Tested with an overridden MediaQuery, should work for iOS.